added cors, created case page
I enabled CORS, setup the overall file structure for the project, created the navbar and added links to Case, Task, and Home Page.

Case's layout is 70% done except for enabling the going into the specific case by clicking on the card's button.

The agents.js file in the api is used to centralize the api endpoint calls, you can see the way I used and customize it for your use case.

For the css I used scss :+1: